Adding some documentation comments for z_zone
and also adding PU_FREE for tags free'd as well as a wider type for sqrt in p_pspr

+++ src/engine/p_pspr.c
@@ -1039,7 +1039,7 @@ void A_FireLaser(player_t* player, pspdef_t* psp) {
 		z = (laser[i]->z1 - laser[i]->z2) >> FRACBITS;
 
 		laser[i]->dist = 0;
-		laser[i]->distmax = (int)sqrt((x * x) + (y * y) + (z * z));
+		laser[i]->distmax = (int64)sqrt((x * x) + (y * y) + (z * z));
 
 		laser[i]->next = NULL;
 		laser[i]->marker = NULL;
